# webui/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "CodeExploder.Gateway | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/sessions/{id}/kb": {
      "get": {
        "tags": [
          "CodeExploder.Gateway"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KbStatus"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/sessions/{id}/threads": {
      "get": {
        "tags": [
          "CodeExploder.Gateway"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ThreadDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "CodeExploder.Gateway"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewThreadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/threads/{threadId}/messages": {
      "get": {
        "tags": [
          "CodeExploder.Gateway"
        ],
        "parameters": [
          {
            "name": "threadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QaMessageDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "CodeExploder.Gateway"
        ],
        "parameters": [
          {
            "name": "threadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewMessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/messages/{messageId}/cancel": {
      "post": {
        "tags": [
          "CodeExploder.Gateway"
        ],
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/analyses/{analysisId}/chunks/{chunkId}": {
      "get": {
        "tags": [
          "CodeExploder.Gateway"
        ],
        "parameters": [
          {
            "name": "analysisId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "chunkId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChunkPeek"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChunkPeek": {
        "required": [
          "path",
          "startLine",
          "endLine",
          "language",
          "content"
        ],
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "startLine": {
            "type": "integer",
            "format": "int32"
          },
          "endLine": {
            "type": "integer",
            "format": "int32"
          },
          "language": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "JsonElement": {},
      "KbStatus": {
        "required": [
          "embeddedChunks",
          "totalChunks",
          "ready"
        ],
        "type": "object",
        "properties": {
          "embeddedChunks": {
            "type": "integer",
            "format": "int32"
          },
          "totalChunks": {
            "type": "integer",
            "format": "int32"
          },
          "ready": {
            "type": "boolean"
          }
        }
      },
...
      "NewMessageRequest": {
        "required": [
          "content",
          "sectionContext"
        ],
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "sectionContext": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
      "NewMessageResponse": {
        "required": [
          "userMessageId",
          "assistantMessageId"
        ],
        "type": "object",
        "properties": {
          "userMessageId": {
            "type": "string",
            "format": "uuid"
          },
          "assistantMessageId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "NewThreadRequest": {
        "required": [
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "QaMessageDto": {
        "required": [
          "id",
          "ord",
          "role",
          "content",
          "status",
          "citations",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ord": {
            "type": "integer",
            "format": "int32"
          },
          "role": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "citations": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JsonElement"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ThreadDto": {
        "required": [
          "id",
          "title",
          "createdAt",
          "lastMessageAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastMessageAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      }
    }
  },
...
}